backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
google-generativeai
psycopg2-binary
requests